[OOT] vfs: Add a trace point in the mark_inode_dirty function
This is an out of tree patch carried by the powertop upstream git
tree.  It is used by powertop to track disk usage.  It is not
upstream because nobody wants to mess with the vfs for debug/analysis.

Signed-off-by:  Kristen Carlson Accardi <kristen@linux.intel.com>

(original commit message below)

PowerTOP would like to be able to show who is keeping the disk
busy by dirtying data. The most logical spot for this is in the vfs
in the mark_inode_dirty() function, doing this on the block level
is not possible because by the time the IO hits the block layer the
guilty party can no longer be found ("kjournald" and "pdflush" are not
useful answers to "who caused this file to be dirty).

The trace point follows the same logic/style as the block_dump code
and pretty much dumps the same data, just not to dmesg (and thus to
/var/log/messages) but via the trace events streams.

Eventually we should be able to phase out the block dump code, but that's
for later on after a transition time.