[OOT] Updated PowerTOP AHCI ALPM accounting for Device Sleep.
This is an out of tree patch, retrieved from the powertop upstream
git tree.  It has been carried out of tree for over 5 years now.
It isn't on a path to upstream because the maintainers objected
to where the statistics were being tracked in sysfs.  A JIRA is
filed to track progress towards upstream, but as it is a P2 item
and hasn't been updated *ever*, I suspect that it isn't going anywhere.
It also has a corner case in which it may not be accurate

https://jira01.devtools.intel.com/browse/LCK-518

Meanwhile, this patch is used by powertop and other tools to
gather SATA LPM statistics and is considered quite useful by our
internal teams.

 --Kristen

This patch applies to v3.7 onwards (including 3.9-rc4), and adds devslp
support to the AHCI state accounting. It's based on the patch at
https://raw.github.com/fenrus75/powertop/master/patches/linux-3.3.0-ahci-alpm-accounting.patch

From: Arjan van de Ven <arjan@linux.intel.com>
Subject: libata: Add ALPM power state accounting to the AHCI driver

PowerTOP wants to be able to show the user how effective the ALPM link
power management is for the user. ALPM is worth around 0.5W on a quiet
link; PowerTOP wants to be able to find cases where the "quiet link" isn't
actually quiet.

This patch adds state accounting functionality to the AHCI driver for
PowerTOP to use.
The parts of the patch are
1) the sysfs logic of exposing the stats for each state in sysfs
2) the basic accounting logic that gets update on link change interrupts
   (or when the user accesses the info from sysfs)
3) a "accounting enable" flag; in order to get the accounting to work,
   the driver needs to get phyrdy interrupts on link status changes.
   Normally and currently this is disabled by the driver when ALPM is
   on (to reduce overhead); when PowerTOP is running this will need
   to be on to get usable statistics... hence the sysfs tunable.

The PowerTOP output currently looks like this:

Recent SATA AHCI link activity statistics
Active	Partial	Slumber	Device name
  0.5%	 99.5%	  0.0%	host0

(work to resolve "host0" to a more human readable name is in progress)

Signed-off-by: Arjan van de Ven <arjan@linux.intel.com>
Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>
Signed-off-by: David E. Box <david.e.box@linux.intel.com>

[Mika and his peers added code to the same place in libahci.c as this
patch did originally-- see bbbc13c799c2 ("ahci: Add functions to manage
runtime PM of AHCI ports"). Move the conflicting hunk down a bit to
prevent a topic branch merge conflict with Mika. 2016-Feb-22]

Signed-off-by: Joe Konno <joe.konno@intel.com>