Add code to create a system validation character device.
Add a modified upstream qib_file_ops.c as file_ops.c.  Most
of the file is commented out.  Added code to create a device
to access the mapped MMRs for diagnostics.  Adjusted the cdev
initialization so we can find the device from the character filp.
Fixed some checkpatch issues with the original.