dynamically obtain the target kernel version
Find the target kerenel version and use it, rather than a hard
coded name.  Allow for a default so the target does not need to
be specified at RPM build time.

Also, rename the ksrc variable (not the directory) to kbuild.
It _is_ a build directory, not a source directory.