wfr-scripts: Fix out of tree build with new Drop 2 changes.
*** NOTE ***: Once the new ioctl and URMPP are accepted upstream we should be
              able to revert this patch.

Because URMPP and the new registration ioctl change the mad_agent structure.
The functions which pass mad_agents around are no longer versioned properly.
For now simply include these modules in the build.  For STL release we should
be able to get the changes upstream and not have an issue with later kernels.