squash: ib/mad: fix loopback; ib_create_send_mad remove base version from qpn
The base version was stashed in the upper 8 bits of the qpn in
ib_create_send_mad to preserve compatibility with other modules.  This needs to
be stripped before the qpn can be passed around locally.  This only affected
local operations because the QPn on the wire is only 24 bits as was truncated
prior to sending...