Fix seg_pio_copy_mid()
A bug was introduced into the seg_pio_copy_mid() helper,
mid_copy_mix(), when the PIO buffer structure internalized
the qw_written and carry values.  The bug is that
mid_copy_mix() can incorrectly update qw_written because
it doesn't add in the doubleword in the carry.  The fix is
to add the doubleword in the carry when calculating
quadwords written.