IB/core: Add cap_opa_mad helper function
OPA MADs share a common header with IBTA MADs but with a different base version
and an extended length.  These MADs increase the performance of management
traffic on OPA devices.

Sharing a common header with IBTA MADs allows us to share most of the MAD
processing code when dealing with OPA MADs in addition to supporting some IBTA
MADs on OPA devices.

Add a port capability flag to indicate OPA MAD support on a port.

Signed-off-by: Ira Weiny <ira.weiny@intel.com>

---
Changes from v4:
	Add cap_opa_mad based on port attr