Add vlflow disable mask option and support
- Even before there are communications with an FM, the default is to be
configured for unreliable transport on VL15 so that directed route packets
behave with same characteristics as they do in InfiniBand.  A new module
parameter sets this default, but also gives a way-out should some other
behavior be desired instead.
- Whenever the link is reset, it goes through linkup_default().
Linkup_default() is where the vpi’s flow_control_mask is now initially set to
the local default.
- When the SM wants to change the configuration, it does so by setting
portInfo. Subn_set_stl_portinfo is now updated to set this information (on the
virtual_stl port, since there is no equivalent in IB).
- There was no need to modify subn_get_stl_portinfo, as the current
implementation already provided the virtual_stl port info that had been set as
a result of the memcpy within that function