Delete rtc-add-an-alarm-disable-quirk.patch (bnc#805740)
This was an initial bug fix from bnc#805740 which we decided not
to bring upstream because it was a brown paper bag. However,
upstream wanted the fix too as someone found the fix in our repo so
John Stultz produced an upstream version which went in into 3.14
(d5a1c7e3fc38d9c7d629e1e47f32f863acbdec3d).

As the upstream version is tagged for stable, I'm dropping our version
here as SLE12 will pick up the stable version.

Signed-off-by: Borislav Petkov <bp@suse.de>

suse-commit: ed8a8549e53b4d8a07f34577010f39f149df3205