IB/hfi: init sc2vl tables to sane defaults
The following values are set:
- SC 0-7   -> VL 0-7 (respectively)
- SC 15    -> VL 15
- SL 8-14  -> VL 0
- SL 16-31 -> VL 0

The hardware powers the tables on to zero, which
prevents VL15 SM packets from working.

The patch adds a new routine, init_sc2vl_tables(),
that is called by the init_chip() routine.  The patch
removes the code that zeros the table values in
reset_txe() and adds a comment noting the move.

Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>