test/LoadModule: Fix missed init of opensmhost and add host names to prints
If opensm is not running when the test is run the host on which to stop it after
starting does not get initilaized and throws an error trying to SSH to None.
Also adds host names to log messages to make it easier to see what host is being
operated on.

Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>