- Update config files: revert LEDS_CLASS to built-in
The core leds module is small (14 kB on x86_64) and many drivers
depend on it or even select it, in particular laptop-specific
drivers and wireless network drivers. So the core leds module
will end up being loaded on most laptop systems, where boot time
matters. On top of that, different architectures have different
dependencies, some of them requiring that the core leds module is
built-in. So the easiest way forward is to just keep the core
leds module built into the kernel on all architectures.

suse-commit: 5f654cf7b4e8c6f0ca0bf0bf1b651d157384d9f9