IB/hfi: Correctly compute context's PIO address
The driver mistakenly used the user context number
instead of the send context number when calculating
the PIO and PIO SOP addresses.
This worked when there was only one kernel context
allocated but when there are multiple ones, the
user context number and the user context's send context
number are not necessarily the same value.

Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>