IB/hfi: Separate RcvArray entries on a per-context basis
One of the performance improvements that WFR offers is the
ability to write a cacheline-sized number of RcvArray
entries all at once to optimize PCIe traffic.

In order to facilitate this, the RcvArray entries need to
divided and handed out to the various receive contexts on
a cacheline boundary. This "forces" the division of RcvArray
entries into groups of 8 at driver load time.

The driver records how many groups each of the receive
contexts is going to be allocated. Any left over groups are
prioritized to user receive contexts. If the number of left
over groups is higher than the number of configured user
contexts, the remainder is divided among the kernel receive
contexts.

This change also imposes a Eager/Expected split limitation on
user level libraries (PSM) - the split between Eager and
Expected RcvArray entry count should also be on a group
boundary.

Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>