IB/hfi: Convert opcode counters to per-context
Profiling has shown that when mulitple contexts are being used there
is a lot of cache activity associated with these counters.

The code formerly kept these counters per port, but only provided the
interface to read per HCA.  This patch converts the read of counters
to per HCA and adds the debugfs hooks to be able to read the file as a
sequence of opcodes.

[Ported from qib]

The qib code has a bug where multiple HCAs (Dual rail) don't
create the second file because timing of the knowing the unit number
and the debugfs stuff is backwards.  This caused the directory
create to fail since the dd->unit isn't correctly initialized.

Change-Id: Ia27ec29da2fd9b215faaabe49947a25e74dd0c0d
Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Roland Dreier <roland@purestorage.com>