IB/qib: Fix memory leak when freeing Eager buffers and status pages
The driver allocats Eager buffer memory in chunks of 32KB. However,
when freeing the memory, it freed only a portion of each chunk due
to the usage of the wrong size (eager buffer size vs. eager buffer
chunk size).

There was also another minor memory leak when the driver did not
free two pages used for communicationg status and events.

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>