IB/hfi: Fix over-mapping of user memory
The rework of the mapping algorithm in commit
8b647b452fa1e97cee824b19ce2afc9f33c3c6cc
introduced a mistake that resulted in the
incorrect over-mapping of user memory due to
wrong calculation of remaining user buffer
to be mapped.
This change fixes that by subtracting the
length of the already mapped user buffer from
the overall length on subsequent iterations
of the mapping loop.

INTERNAL: PR 124102 - Expected receive support in Driver with
INTERNAL: Functional Simulator

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>