ib/mad: STL handle variable length MAD's (partial some send support already existed)
Jumbo MAD's can be shorter than 2048.  Support this processing by:

* Using the RQ WC.byte_len when processing received MADs
* Adjusting the 'fake' WC to represent the proper incoming byte_len directed to
  the local SMA
* Use the response length indicated from the local HCA agents
	1) when sending agent responses on the wire
	2) when passing responses through the local_completions function

NOTE: GRH length is specified to the local HCA agents but is not passed to the
registered mad agents.  This is in keeping with the current lengths which are
specified for IB MAD's.

INTERNAL: PR 124154 - oib_utils api: oib_send_recv_mad_no_alloc does not indicate rx size

Change-Id: Ibe403fc107dbb5733c9220b320c1a66239a420e9
Signed-off-by: Ira Weiny <ira.weiny@intel.com>
Reviewed-by: John Fleck <john.fleck@intel.com>
Reviewed-by: Anton Bodner <anton.bodner@intel.com>