IB/hfi: add debugfs hooks for hfistats counter files
These files are moved from the problematic ipathfs
file system to debugfs.

The current seq_file oriented debugfs is generalized
to support non seq_files to ease the migration of
the fs.c handling.

As part of the move, the per port readers now
use a generalized routine with the file private
data having the ppd pointer.  This avoids the ugly
routine for each port hack in the fs.c

As of this patch, all handling in fs.c has moved
to debugfs.c.

INTERNAL: PR 124131 - Add kernel hooks for hfistats

Change-Id: Iab846ca2f947406bd20d812243f01562d1301d43
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>