IB/hfi: add counter/stats interface to wfr
This patch adds the counter interface for hfistats
to be read by the wfr port of ipathstats.  The
implementation is as described in the WFR driver
design specification.

The significant difference between qib and wfr
is that the interfaces have moved from /ipathfs
to debugfs, avoiding issues of portablity and
reliability with having a driver private filesystem.

The stats/counter interface is entirely table driven
with a flexible entry that can be extended to for
synthethic counters per HAS 13.2. (As of now,
none of these counters are implemented.)

There is a table for device and port counters.  In many
respects, with a single port design, the choice of
what stat is device relative or port relative is arbitrary.
In this case, the choice was made based on SusieQ
positioning of the equivalent stat.

Other notable differences:
- fs.c is deleted including changes for getting
  rid of calls
- The counter defines are added in such as way as
  to be upward compatible when those defines are
  eventually autogenerated
- The stats timer is now consolidated with these
  changes instead of being spread out all over the
  driver
- The unused struct qlogic_ib_counters is deleted

INTERNAL: PR 124131 - Add kernel hooks for hfistats

Change-Id: I34ead4bd319587c4bdacb9858f45cf20a91c700b
Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>