Initial test harness fixes for FPGA support in loopback mode
This removes the explicit exit when not using --simics. The patch also
respects whether --simics is used or not when running through the harness and
propagates that through to the individual tests.

The old SLES install that is currently being used on the FPGA host does not
have one of the required git commands so this patch handles that gracefully.

INTERNAL: PR 124276 - the driver test harness does not work on fpga1

Change-Id: I7bfd0855e0b66e88c539b8a6970c21c2857ad2d5
Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>