Regression tests for building and installing RPMs and testing hfi_control
This test will build and install the given RPMs on your host.
Note if doing this with simics the changes will not be persistent unless you
create a craff. The test script for hfi_control is also included. These are
both being done to enable testing of hfi_control.

INTERNAL: PR 124091 - WFR Support for tracing and hfi_control

Change-Id: I0bd47de5bb796d5a7adc19abc8d1c2b9fa021a2a
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>