IB/hfi: fix install dependency
The current install fails because of a
dependency issue. A --nodeps was being
used to detour the issue.

This is due to the fact that the base kernel
provides 3.xx.xx_wfr+ vs. 3.xx.xx-wfr+ as
in the hfi.spec.in.

This fix translates the %kver '-' to an '_' to
correct the issue.

INTERNAL: PR 124146 - wfr-driver has rpm install dependency problems

Reviewed-by: John Fleck <john.fleck@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>