IB/hfi: fix pkey change event port
The code used a literal 1 in dispatching
an IB_EVENT_PKEY_CHANGE.

Since WFR only supports 1 port, this patch
isn't really all that important, but
the mad code shouldn't make that assumption.

Change to use the port as specified in the call.

INTERNAL: PR 123954 - IPoIB fails to work with default config

Change-Id: Iec9ada16d87493718049f6df7a800a9065b1afa1
Cc: <stable@vger.kernel.org>
Reported-by: Alex Estrin <alex.estrin@intel.com>
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>