supported.conf: Remove outdated entries
Remove all modules which no longer exist, or which we no longer
build. I have carefully reviewed each of them to prevent any false
positive, so hopefully this big cleanup won't break anything.

Note: modular code which we decided to build into the kernel was not
removed, on purpose, in case we change our mind later for whatever
reason.

suse-commit: d04f46d32d61ff01d623c784e922f4fdf6c135f8