IB/hfi: Change scope of P_Keys in driver structures
P_Keys are only meaningful on a port level and the
Infiniband fabric has no notion of contexts. Pior to
this change, the primary container for the programmed
P_Key table was the per-context structure.

This might have made sense for previous hardware since
the first two kernel contexts were hard-wired to the
physical ports. However, for this version of the hardware
there is no such association. Therefore, the proper
place to store all programmed P_Keys becomes the per-
port structure.

Another change that applies to the new HW is that the
FM is responsible for managing the PKey table. This
means that we can accept the PKey table as is and
store it index-by-index into our internal structure.

The only exception to the above rule is that we have
to ensure that, at a minimum, the partial membership
PKey is present in the table. This check is already
being performed.

This change also adds the calls to program the HW
PKey table.

INTERNAL: PR 124194 - driver needs to support PSM picking
INTERNAL: a p_key for its user send context

Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I418dd7434c8f2ef9802832dc6f3f577539450e43