IB/hfi: Add support for J_KEY
The purpose of Job Keys is to prevent jobs from
sending packets to node/processes not part of the
job.

The device driver will generate a JobKey for each
job (currently, it is the lower 16bits of the user's
UID) and program Hardware registers (both TXE and RXE)
with that value.

A new function is introduced to generate the JobKey
allowing for the driver to use a common generation
method for context used by both kernel and user ULPs.

INTERNAL: PR 124103 - [#4] Add J_KEY support in Driver

Reviewed-by: Dean Luick <dean.luick@intel.com>
Reviewed-by: John Gregor <john.a.gregor@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: Id880af2e8218581376e63bf0b549f541ab778bc6