IB/hfi: Add support for per-VL MTUs and STL MTUs
The STL architecture provides a per-VL MTU, unlike
standard IB. Therefore, the MTU is not a "global"
size for the port. The FM can chose to set different
MTU values for each of the configured VLs.

The driver needs to start keeping track of the
configured MTUs (FM will supply the configuration
through STL extended MADs) and configure each HW
VL SendLength accordingly.

Since we need to have a per-VL data structure for
the MTUs, use the same structure to hold other
per-VL data (currently, the per-VL dedicated send
contexts).

In order to lessen the work that PSM has to do in
case that it does not perform the path record
query, the driver also exports the programmed per-VL
MTUs through SYSFS. PSM can look up the MTU for a
VL after it has resolved the SL2VL mapping.

INTERNAL: PR 124218 - Support all MTU supported by
INTERNAL: STL1 in Driver

Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I8d2b7c7743009038088228cc0e0c1db4198c5f27