IB/hfi: implement processing of buffer control tables
This patch implements the buffer control table
get/set functionality to support the STL FM.

The existing code had the query functionaliy
but the processing didn't match the big endian
format of the STL messages.

This patch adds the processing in the SMA
as well as adjusting the internal processing
to operate in big endian.

INTERNAL: PR 124107 - STL Enhanced MAD support in driver

Change-Id: I0fff54356551369eeaf561f860e7cd4f44eba8cd
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>