Ib/hfi: Fix setting of Link state flags
PSM uses the link state flags mapped into it's
memory space to check what the status of the
link is when an error occurs.

The move to the new WFR link state machine and code
broke the setting of the flags by setting the
IB_READY flag on Physical LinkUp instead of logical
LINK_ACTIVE.

This change unifies the setting of the flags used
by user space to a single function and corrects the
setting of the bits depending on the link state.

INTERNAL: PR 124093 - WFR Host SW Drop 4

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>