IB/HFI: Snoop/capture support.
This patch adds snoop and capture support to HFI. It differs significantly
from the qib version. Most notably there is a single device for both snoop and
capture, the file open mode determines the mode of operation. The code is also
architected out of the send and recv paths and uses function pointers to do the
right thing, so we avoid all the conditional logic.

INTERNAL: PR 123996  - WFR support for packet capture and packet snoop

Change-Id: I1225dfc98ffaad4a8a53577b31cf222c662692a6
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>