IB/hfi: Rework code that sets PSM link state flags
The initial patch which attempted to fix the setting
of the link state flags for PSM was incomplete as it
did not handle all cases - there were code paths which
change the logical link state but did not go through
qib_set_linkstate(). Therefore, the flags were not
always updated.

Moving the code which sets the flags to the function
which also updates the logical link state cached value
fixes the omitted cases because the cache update
function is called by all code paths.

INTERNAL: PR 124093 - WFR Host SW Drop 4

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: Ieaa34524ae351d634cab4b9c692c7f6e0ed585c7