IB/hfi: Set max MTU for the port in DC
The DC needs to be programmed with the maximum MTU for the
port (it defaults to 4K). Otherwise, packets will be dropped
on receive if their length is greater than the DC setting.

This prevented anything from using 8K MTU successfully.

INTERNAL: PR 124457 - DC badpktlen errors for > 4KB messages
INTERNAL: with wfr-psm with OsuMpi.py reg test

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I4dea7af363f1a27663daa448172ab7695c75d631