Update ppc64le config files: Disable useless drivers
SLE12 on ppc64le aims server hardware only, so disable a bunch of
drivers which are irrelevant there:

* gameport and joystick drivers
* tablet drivers
* touchscreen drivers
* misc input drivers
* memorystick drivers
* sound drivers
* media drivers

I also disabled 5 legacy PATA drivers, not needed on POWER8 hardware.

This puts the ppc64le config more in line with the s390x config, which
aims at the same hardware class.

suse-commit: 7e03116c15c8acf96845e45bd5986afc3aff007f