IB/hfi: Do not watch LCB errors when assigned to 8051
To prevent conflicts, the host driver must stop watching the
LCB errors when it assigns LCB access to the 8051.  Most LCB
errors are informative only.  The 8051 watchs the fatal LCB
errors and will inform the host with a link down state change.
The host can at that point examine the register, if it wishes.

INTERNAL: PR 124220 - FPGA_P power on and functional

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>