IB/hfi: Add OSU benchmark paths for FPGA
The OSU benchmarks are not in the same location on the FPGA as
they are in simulation.  Change how the path is obtained.

o OsuMpi.py: Fix the test module name.
o OsuMpi.py: Add no prefixes to the test names.
o OsuMpi.py: When constructing the ssh command, do not add hard-coded
  path parts, rearrange so that each line adds a complete option and
  argument pair.
o RegLib.py: Change get_osu_benchmark_dir() to look for the OSU
  directory on the first test node.

INTERNAL: PR 124220 - FPGA_P power on and functional

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>