IB/hfi: Fix overflow error for link speed and return max of EDR speed.
This patch fixes an overflow bug causing a 0 value to be returned in port
info query. What was happening was a 16bit value was being saved into an 8 bit
value so the upper bits were getting lost.

The other issue is that MPI supports only up to EDR. So this fix knocks the
maximum speed down to EDR.

INTERNAL: PR 124517 - MPI with verbs regression test failure

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: I6e6bd1671c5bec24689fce314bc504f3c947d7c5