Hfi/Tests: Add script to handle mounting a craff on a simulation host
This script will uncompress a craff, mount it, then change root into said
craff. Once the user is done working with the craff they type exit and the
craff is then saved, and the uncompressed image deleted, or the craff if not
saved and the uncompressed image is preserved. The behavior is dictated by if
an optional output craff file is given on the CLI or not.

INTERNAL: PR 124240 - Put mpi_stress into the driver regression tests.

Reviewed-by: Andrew Friedley <andrew.friedley@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: I1877ded2a8a81bb77f81f844e5e961a24155ec2c