IB/hfi: Add support for user level PKeys
Send-side PKey checking requires the
context user to set the appropriate PKey
for the context.

To do this, the driver provides this new
command that will program the desired
PKey into the contexts CSRs and turn send
PKey checking on.

In order to prevent arbitrary PKeys from
being set by user-level libraries and
applications, the input PKey is checked
against the PKey table provided by the SM.
If the input PKey does not match any of the
key programmed by the SM, the command is
rejected. The driver also ensures that the
input PKey isn't one of the management PKeys
as Omni Scale does not allow data traffic
on management partitions.

Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>