Honor --sw-diags in HfiPktSend.py and HfiPktTest.py
Use the --sw-diags parameter to determine where the pkt_test/send tools
are.  By default, the tests assume the tools are in the default path.
If --sw-diags is specified, it is interpreted as a path to a
wfr-diagtools-sw source tree with the tools compiled in place.

INTERNAL: PR 124625 - HfiPktTest fails when not installed in craff

Reviewed-by: Denny Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Andrew Friedley <andrew.friedley@intel.com>