IB/hfi: Fix programing of RcvKeyCtrl CSR
While programing the RcvKeyCtrl CSR, the
driver was calling write_uctxt_csr() instead of
write_kctxt_csr(). Due to the different CSR
strides for the two types of functions/CSRs,
calling the wrong function resulted in programming
the CSR for the incorrect context.

Eventually, this would lead to writing of a bad
address when the receive context number got higher
than 10.

INTERNAL: PR 124702 - WFR driver writes RcvKeyCtr
INTERNAL: to the wrong places

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I53da0593646b3ec167a046563b410f783a9403cd