IB/qib: Port qib tracing feature from IFS 7.2.2
This patch adds tracing feature to qib driver and the
qib_log/qib_cdbg calls that were part of IFS 7.2.2.

This patch will allow for compilation when _QIB_DEBUGGING
is not defined and fixes many checkpatch issues in
the IFS 7.2.2 code base.

INTERNAL: PR 124506 - ipath_trace command is not working with IFS-7.3

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Vinit Agnihotri <vinit.abhay.agnihotri@intel.com>