IB/hfi: Prevent out-of-bounds RcvArray programming
The driver had a bug which would cause it to use
RcvArray entries for Expected receives that were
past the allocated range for the context.
This was due to a missing check that would ensure
that the number of remaining free groups tracked
by the group bit mask was bound to the number of
groups allocated to the context.

This bug only happened under certain conditions -
when the number of allocated RcvArray groups for a
context did not fit nicely into a bit vector.

This commit also adds a performance improvement to
the Expected receive code by creating and using a
dedicated tracepoint.

INTERNAL: PR 124798 - WFR driver incorrectly programs
INTERNAL: expected RcvArray entries

Reviewed-by: John Gregor <john.a.gregor@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I2f67c430fd91715da339ec35db5de69199553c3e