IB/hfi: fix errors with for_each_set_bit() usage
1) for_each_set_bit() is being used incorrectly - the last
argument should be a number of bits (not bytes) to scan.

2) generally the per-VL MAD response data is "packed", e.g.,
if the vl_select_mask is 0x8001, then there are only 2 per-VL
data sections, one for VL0, and one for VL15, in a 2 entry
array. The code used to use the bit# to index into the per-VL
array, which is obviously wrong. (The format of per-VL MAD
data is discussed in the STL Gen 1 spec., Vol. 1, section
20.4.2.3 (and subsections)).

INTERNAL: PR 124667 - VL15 PMA query indicates no data

Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: I46a8503bbae5dbe434000a18448ab8cd06500d21