IB/hfi: Add "#ifdef CONFIG_STL_MGMT ... #else ... #endif" blocks throughout mad.c to separate code that's used by IFS FM, and opensm.
I'm about to submit a patch which will modify struct qib_ibport
so that the structure's content will depend on whether or not
CONFIG_STL_MGMT is defined. Any code that uses the affected structure
elements will have to be within #ifdef CONFIG_STL_MGMT ... #else ...
#endif" blocks. Practically, this means that much (not all) of mad.c
is now compiled conditionally on the value of CONFIG_STL_MGMT, and
most importantly, that the driver will work with opensm iff
CONFIG_STL_MGMT is _not_ defined, and it will work with the IFS FM
iff CONFIG_STL_MGMT _is_ defined.

Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: Ia824cb7d2c1fd6ec6f5e61e00422645c97c799a1