IB/hfi: Implement PSM SDMA API
This commit includes the PSM SDMA API and all
related code. The PSM SDMA code is based on the
common SDMA API that is also used for Verbs and
provides a common way to access and use the SDMA
engines.

At this stage, the PSM API is semi-blocking - the
system call into the driver blocks the PSM process
until all packets for the message have been submitted
to the SDMA engine. However, it does not wait for
send completions. The completions are done
asynchronously and the PSM code signals the PSM process
of the completion of an entire message.

This commit also includes code for a more non-blocking
implementation, which uses a kernel thread to submit
the remained of the message packets after an initial set
is submitted by the system call. However, at this stage
this kernel thread is not used and the corresponding
code is compiled out.

INTERNAL: PR 124935 - Implement PSM use of common SDMA API
INTERNAL: PR 124342 - Add PSM interface to use SDMA via driver

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: John Gregor <john.a.gregor@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I4752f90183b0ca7235dbb9458644440071a52121