IB/hfi: Support metadata for captured packets in regression suite scripts.
A metadata header is being added to captured packets, this patch updates the
regression tests to handle it. Specifically PcapLocal.py. There are changes
required for Snoop tests as well, to account for the metadata when using
packet capture mode.

INTERNAL: PR 124426 -  Add metadata to the Capture/Snoop port that can be
INTERNAL: used by iba_packet_capture/fabric_sim

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: Ifad864e6f2a4f0637cc700a4e757dcc318bbedcc